Refactor code for creating a new post in HTML template
This commit refactors the code for creating a new post in the HTML template. It adds variables to retrieve form inputs using query selectors and assigns their corresponding values to the post object. The tags are split from a string input and stored in an array. The post object is then formatted with the retrieved values and stored in the appropriate variables.