Add benchmarking script to measure response time
This commit adds a new script called benchmark.ps1 that measures the response time of a website by making multiple HTTP requests. The user is prompted to enter the URL of the website and the number of requests to be made. The script validates the URL format and the input for the number of requests. It then makes the requests and collects the response times. Finally, it calculates and outputs the average and maximum response times.