feat: Add functionality for creating new pages
This commit adds the functionality for creating new pages. It includes a form where users can add data types for each component they want to add to the page. After adding the data types, users can submit the form to create the page. The selected data types are then displayed in a table below the form.