Update dependencies: remove unused library and upgrade existing libraries in Cargo.toml.
The commit updates dependencies in `Cargo.toml`, removing the unused library `mediasoup` and upgrading existing libraries `strum` and `strum_macros` to versions 0.25.0 and 0.25.1 respectively. The library `rml_rtmp` is also upgraded to version 0.8.0, and the `uuid` library is updated to version 1.4.1, including the feature "v4". The `sqlx` dependency is modified to include features "postgres", "macros", "chrono", "json", "uuid", and "offline".